made so many changes, tracking goes out of hand
more monkey patches

+++ Assets/_Project/Scripts/AbilitySystem/Abilities/DoubleJumpAbility.cs
@@ -20,7 +20,7 @@ public class DoubleJumpAbility : Ability
     public override bool IsTriggered()
     {
         return heroController.IsGrounded == false
-            && abilityUserBlackboard.GetState((int)InputTypeId.Jump);
+            && abilityUserBlackboard.GetState(InputTypeId.Jump);
     }
 
     public override bool CanBeUsed()
@@ -43,9 +43,9 @@ public class DoubleJumpAbility : Ability
 
     public override void FixedUpdate()
     {
-        float direction = abilityUserBlackboard.GetAxis2D((int)InputTypeId.Move).x;
+        float direction = abilityUserBlackboard.GetAxis2D(InputTypeId.Move).x;
 
-        float speed = abilityUserBlackboard.GetState((int)InputTypeId.Run)
+        float speed = abilityUserBlackboard.GetState(InputTypeId.Run)
             ? dataService.HeroData.RunSpeed
             : dataService.HeroData.MovementSpeed;
