+++ Assets/_Project/Scripts/AbilitySystem/Abilities/FallAbility.cs
@@ -29,9 +29,9 @@ public class FallAbility : Ability
 
     public override void FixedUpdate()
     {
-        float direction = abilityUserBlackboard.GetAxis2D((int)InputTypeId.Move).x;
+        float direction = abilityUserBlackboard.GetAxis2D(InputTypeId.Move).x;
 
-        float speed = abilityUserBlackboard.GetState((int)InputTypeId.Run)
+        float speed = abilityUserBlackboard.GetState(InputTypeId.Run)
             ? dataService.HeroData.RunSpeed
             : dataService.HeroData.MovementSpeed;
