+++ Assets/_Project/Scripts/AbilitySystem/Abilities/JumpAbility.cs
@@ -19,7 +19,7 @@ public class JumpAbility : Ability
 
     public override bool IsTriggered()
     {
-        return abilityUserBlackboard.GetState((int)InputTypeId.Jump);
+        return abilityUserBlackboard.GetState(InputTypeId.Jump);
     }
 
     public override bool CanBeUsed()
@@ -44,9 +44,9 @@ public class JumpAbility : Ability
 
     public override void FixedUpdate()
     {
-        float direction = abilityUserBlackboard.GetAxis2D((int)InputTypeId.Move).x;
+        float direction = abilityUserBlackboard.GetAxis2D(InputTypeId.Move).x;
 
-        float speed = abilityUserBlackboard.GetState((int)InputTypeId.Run)
+        float speed = abilityUserBlackboard.GetState(InputTypeId.Run)
             ? dataService.HeroData.RunSpeed
             : dataService.HeroData.MovementSpeed;
